[Gold V] Title: 시간 구간 다중 업데이트 다중 합, Time: 660 ms, Memory: 116848 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/25827. 시간 구간 다중 업데이트 다중 합/시간 구간 다중 업데이트 다중 합.java
@@ -10,7 +10,7 @@ public class Main {
         
         int query = Integer.parseInt(br.readLine());
         
-        int[] diff = new int[SEC+1];
+        long[] diff = new long[SEC+1];
         
         StringTokenizer st;
